pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "driveratlas"
version = "0.1.0"
description = "Windows kernel driver structural analysis toolkit"
requires-python = ">=3.10"
dependencies = [
    "pefile>=2023.2.7",
    "click>=8.0",
    "rich>=13.0",
    "PyYAML>=6.0",
    "cryptography>=41.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
driveratlas = "driveratlas.cli:main"

[tool.setuptools.packages.find]
include = ["driveratlas*"]
